db/update/Walk:FindAncestorLoop: uint64_t inode & device
Previously, inode numbers were truncated to 32 bits, which could lead
to problems on XFS where inodes are 64 bit; this could lead to bogus
"recursive directory found" errors during database update.

[mk: added commit description and NEWS line]

Closes https://github.com/MusicPlayerDaemon/MPD/issues/2000

+++ src/db/update/Walk.cxx
@@ -148,7 +148,7 @@ update_directory_stat(Storage &storage, Directory &directory) noexcept
  */
 static int
 FindAncestorLoop(Storage &storage, Directory *parent,
-		 unsigned inode, unsigned device) noexcept
+		 uint64_t inode, uint64_t device) noexcept
 {
 #ifndef _WIN32
 	if (device == 0 && inode == 0)
